# Encapsulate ordering and naming conventions in separate classes