… clifford/_multivector.py
… clifford/test/test_clifford.py
… clifford/test/test_layout.py
… clifford/test/test_layout_helpers.py
… docs/predefined-algebras.rst