The end goal here is to be able to more clearly separate which bits of `Layout` are specific to the algebra, and which are specific to the choice of blade names and numbers.

This makes the observation that our blade tuples, eg `(1, 2, 3)` &rarr; `e123` need not be integers at all. We already provide some flexibility via `firstIdx`, but the existence of this field resulted in some weird code that was cleaned up in #277. With this change, it's possible (but perhaps not adviseable) to have `('x', 'y', 'z')` &rarr; `exyz`:
```python
In [1]: from clifford import Layout

In [2]: from clifford._layout import BasisBladeOrder, BasisVectorIds

In [3]: l = Layout([1, 1, 1], ids=BasisVectorIds(['x', 'y', 'z']), order=BasisBladeOrder.shortlex(3))

In [4]: locals().update(l.blades)

In [5]: p = ex ^ ey; p
Out[5]: (1^exy)

In [6]: p['x', 'y']
Out[6]: 1

In [7]: l
Out[7]:
Layout([1, 1, 1],
       ids=BasisVectorIds(['x', 'y', 'z']),
       order=BasisBladeOrder.shortlex(3),
       names=['', 'ex', 'ey', 'ez', 'exy', 'exz', 'eyz', 'exyz'])
```

Perhaps more usefully, it's now legal to write just
```python
In [1]: from clifford import Layout

In [2]: Layout([1, 1, 1])
Out[2]:
Layout([1, 1, 1],
       ids=BasisVectorIds(['x', 'y', 'z']),
       order=BasisBladeOrder.shortlex(3),
       names=['', 'e1', 'e2', 'e3', 'e12', 'e13', 'e23', 'e123'])
```

This partially addresses #54.

The size of this patch is inflated a lot by docs and tests - the actual change is perhaps a net +300 line delta.