@eric-wieser 2020-03-10
Failure is real, turned out I made the following changes to the blade ordering:
```
 (1.0^e5)
 (1.0^e12)
 (1.0^e13)
+(1.0^e14)
+(1.0^e15)
 (1.0^e23)
-(1.0^e14)
 (1.0^e24)
+(1.0^e25)
 (1.0^e34)
-(1.0^e15)
-(1.0^e25)
 (1.0^e35)
 (1.0^e45)
 (1.0^e123)
 (1.0^e124)
+(1.0^e125)
 (1.0^e134)
+(1.0^e135)
+(1.0^e145)
 (1.0^e234)
-(1.0^e125)
-(1.0^e135)
 (1.0^e235)
-(1.0^e145)
 (1.0^e245)
 (1.0^e345)
 (1.0^e1234)
```