@eric-wieser 2020-03-10
Performance is pretty untouched:
After:
```
In [3]: from clifford import Cl

In [4]: %timeit Cl(8)
142 ms ± 14.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %timeit Cl(3)
8.75 ms ± 823 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
Before:
```
In [1]: from clifford import Cl

In [2]: %timeit Cl(8)
139 ms ± 8.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [3]: %timeit Cl(3)
9.26 ms ± 1.86 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```